+++ flake8_pyi/errors.py
@@ -127,6 +127,8 @@ Y066 = (
     'put the code for new Python versions first, e.g. "{new_syntax}"'
 )
 Y067 = 'Y067 Use "=None" instead of "Incomplete | None = None"'
+Y068 = 'Y068 Do not use "@override" in stub files.'
+
 Y090 = (
     'Y090 "{original}" means '
     '"a tuple of length 1, in which the sole element is of type {typ!r}". '